requirements.txt
aniso8601==1.3.0
click==6.7
Flask==0.12.2
flask-restplus==0.10.1
Flask-SSLify==0.1.5
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
py-postgresql==1.2.1
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
Werkzeug==0.12.2
